# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Alpha Bit Gadget Shop API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List products with optional filters
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
...
  /products/offers:
    get:
      operationId: getOfferProducts
      tags: [products]
      summary: Get products on sale
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
      responses:
        "200":
          description: Offer products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found

  # ── CATEGORIES ────────────────────────────────────────────────────────────
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a new order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderCreated"
        "400":
          description: Bad request
...
  /admin/stats:
    get:
      operationId: getAdminStats
      tags: [admin]
      summary: Dashboard stats (orders, revenue, customers)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStats"
  /admin/orders:
    get:
      operationId: listAdminOrders
      tags: [admin]
      summary: List all orders
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOrderList"
  /admin/orders/{orderId}/status:
    put:
      operationId: updateOrderStatus
      tags: [admin]
      summary: Update order status
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: orderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /admin/customers:
    get:
      operationId: listAdminCustomers
      tags: [admin]
      summary: List all customers
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerList"
  /admin/crm/stats:
    get:
      operationId: getCrmStats
      tags: [admin]
      summary: CRM stats (top customers, recent sign-ups)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: CRM stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CrmStats"
  /admin/products:
    get:
      operationId: listAdminProducts
      tags: [admin]
      summary: List products (admin)
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: category
          schema:
            type: string
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 100
        - in: query
          name: skip
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminProductList"
    post:
      operationId: createAdminProduct
      tags: [admin]
      summary: Create product
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /admin/products/{productId}:
    put:
      operationId: updateAdminProduct
      tags: [admin]
      summary: Update product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteAdminProduct
      tags: [admin]
      summary: Delete product
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: productId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
  /admin/offers:
    get:
      operationId: listAdminOffers
      tags: [admin]
      summary: List coupons/offers
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Offers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Offer"
    post:
      operationId: createAdminOffer
      tags: [admin]
      summary: Create coupon/offer
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
  /admin/offers/{offerId}:
    put:
      operationId: updateAdminOffer
      tags: [admin]
      summary: Update coupon/offer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: offerId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OfferUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Offer"
    delete:
      operationId: deleteAdminOffer
      tags: [admin]
      summary: Delete coupon/offer
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: offerId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResult"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Product:
      type: object
      required: [id, name, description, short_description, price, category, image, stock, sku]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        short_description:
          type: string
        price:
          type: number
        sale_price:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: string
        stock:
          type: integer
        sku:
          type: string
        featured:
          type: boolean
        on_sale:
          type: boolean
    ProductList:
      type: object
      required: [products, total, limit, skip]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        limit:
          type: integer
        skip:
          type: integer
    ProductInput:
      type: object
      required: [name, description, price, category]
      properties:
        name:
          type: string
        description:
          type: string
        short_description:
          type: string
        price:
          type: number
        sale_price:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: ["string", "null"]
        stock:
          type: integer
          default: 100
        sku:
          type: string
        featured:
          type: boolean
        on_sale:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        short_description:
          type: string
        price:
          type: number
        sale_price:
          type: ["number", "null"]
        category:
          type: string
        image:
          type: ["string", "null"]
        stock:
          type: integer
        featured:
          type: boolean
        on_sale:
          type: boolean
    AdminProductList:
      type: object
      required: [products, categories, total]
      properties:
        products:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        categories:
          type: array
          items:
            type: string
        total:
          type: integer

    # ── Category ──────────────────────────────────────────────────────────────
...
    CartItem:
      type: object
      required: [product_id, quantity]
      properties:
        product_id:
          type: string
        quantity:
          type: integer
...
    CartLineItem:
      type: object
      required: [product_id, name, price, quantity, total]
      properties:
        product_id:
          type: string
        name:
          type: string
        image:
          type: string
        price:
          type: number
        quantity:
          type: integer
        total:
          type: number
        sku:
          type: string
...
    OrderInput:
      type: object
      required: [cart_items, customer_name, customer_email, customer_phone, shipping_address, shipping_city, shipping_zip]
      properties:
        cart_items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        customer_name:
          type: string
        customer_email:
          type: string
        customer_phone:
          type: string
        shipping_address:
          type: string
        shipping_city:
          type: string
        shipping_zip:
          type: string
        coupon_code:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
    OrderCreated:
      type: object
      required: [order_id, total]
      properties:
        order_id:
          type: string
        total:
          type: number
        message:
          type: string
    Order:
      type: object
      required: [id, customer_name, customer_email, status, total, created_at]
      properties:
        id:
          type: string
        customer_name:
          type: string
        customer_email:
          type: string
        customer_phone:
          type: string
        shipping_address:
          type: string
        shipping_city:
          type: string
        shipping_zip:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartLineItem"
        subtotal:
          type: number
        shipping:
          type: number
        discount:
          type: number
        total:
          type: number
        status:
          type: string
        payment_status:
          type: string
        payment_id:
          type: ["string", "null"]
        tracking_number:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        created_at:
          type: string
        updated_at:
          type: string
    AdminOrderList:
      type: object
      required: [orders, total]
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
    OrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
        tracking_number:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Payment ──────────────────────────────────────────────────────────────
...
    CustomerList:
      type: object
      required: [customers, total]
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/CustomerSummary"
        total:
          type: integer
    CustomerSummary:
      type: object
      required: [id, name, email, order_count, total_spent]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        order_count:
          type: integer
        total_spent:
          type: number
        created_at:
          type: string
        tags:
          type: array
          items:
            type: string
    CrmStats:
      type: object
      required: [total_customers, new_this_month, top_customers]
      properties:
        total_customers:
          type: integer
        new_this_month:
          type: integer
        top_customers:
          type: array
          items:
            $ref: "#/components/schemas/CustomerSummary"

    # ── Offer / Coupon ────────────────────────────────────────────────────────
    Offer:
      type: object
      required: [id, code, discount_type, discount_value, active]
      properties:
        id:
          type: string
        code:
          type: string
        description:
          type: string
        discount_type:
          type: string
        discount_value:
          type: number
        min_order:
          type: ["number", "null"]
        max_uses:
          type: ["integer", "null"]
        uses_count:
          type: integer
        active:
          type: boolean
        expires_at:
          type: ["string", "null"]
    OfferInput:
      type: object
      required: [code, discount_type, discount_value]
      properties:
        code:
          type: string
        description:
          type: string
        discount_type:
          type: string
        discount_value:
          type: number
        min_order:
          type: ["number", "null"]
        max_uses:
          type: ["integer", "null"]
        active:
          type: boolean
        expires_at:
          type: ["string", "null"]
    OfferUpdate:
      type: object
      properties:
        code:
          type: string
        description:
          type: string
        discount_type:
          type: string
        discount_value:
          type: number
        min_order:
          type: ["number", "null"]
        max_uses:
          type: ["integer", "null"]
        active:
          type: boolean
        expires_at:
          type: ["string", "null"]
...
    AdminStats:
      type: object
      required: [total_orders, total_revenue, total_customers, total_products]
      properties:
        total_orders:
          type: integer
        total_revenue:
          type: number
        total_customers:
          type: integer
        total_products:
          type: integer
        pending_orders:
          type: integer
        paid_orders:
          type: integer
        shipped_orders:
          type: integer
        revenue_this_month:
          type: number
        orders_this_month:
          type: integer
        recent_orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"

    # ── Contact ───────────────────────────────────────────────────────────────
...
    MessageResult:
      type: object
      required: [message]
      properties:
        message:
          type: string