# lib/api-zod/src/generated/types/customerList.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Bit Gadget Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { CustomerSummary } from './customerSummary';

export interface CustomerList {
  customers: CustomerSummary[];
  total: number;
}