# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Bit Gadget Shop API
 * OpenAPI spec version: 0.1.0
 */
import type { CartLineItem } from './cartLineItem';

export interface Order {
  id: string;
  customer_name: string;
  customer_email: string;
  customer_phone?: string;
  shipping_address?: string;
  shipping_city?: string;
  shipping_zip?: string;
  items?: CartLineItem[];
  subtotal?: number;
  shipping?: number;
  discount?: number;
  total: number;
  status: string;
  payment_status?: string;
  /** @nullable */
  payment_id?: string | null;
  /** @nullable */
  tracking_number?: string | null;
  /** @nullable */
  notes?: string | null;
  created_at: string;
  updated_at?: string;
}