# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Bit Gadget Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: string;
  name: string;
  description: string;
  short_description: string;
  price: number;
  /** @nullable */
  sale_price?: number | null;
  category: string;
  image: string;
  stock: number;
  sku: string;
  featured?: boolean;
  on_sale?: boolean;
}