# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Alpha Bit Gadget Shop API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  description: string;
  short_description?: string;
  price: number;
  /** @nullable */
  sale_price?: number | null;
  category: string;
  /** @nullable */
  image?: string | null;
  stock?: number;
  sku?: string;
  featured?: boolean;
  on_sale?: boolean;
}